Conflicts:
	CMGTools/TTHAnalysis/cfg/run_susyAlphaT_cfg.py

+++ CMGTools/Production/scripts/pybatch.py
@@ -35,45 +35,106 @@ def split(comps):
 def batchScriptCERN( jobDir, remoteDir=''):
    '''prepare the LSF version of the batch script, to run on LSF'''
    
+   dirCopy = """echo 'sending the logs back'  # will send also root files if copy failed
+cp -r Loop/* $LS_SUBCWD
+if [ $? -ne 0 ]; then
+   echo 'ERROR: problem copying job directory back'
+else
+   echo 'job directory copy succeeded'
+fi"""
    if remoteDir=='':
-      cpCmd="""echo 'sending the job directory back'
-cp -r Loop/* $LS_SUBCWD"""
+      cpCmd=dirCopy
    elif remoteDir.startswith("/pnfs/psi.ch"):
        cpCmd="""echo 'sending root files to remote dir'
-export LD_LIBRARY_PATH=/usr/lib64:$LD_LIBRARY_PATH # Fabio's workaround to fix gfal-tools with CMSSW
-for f in Loop/tree*/*.root
-do
-   ff=`basename $f | cut -d . -f 1`
-   d=`echo $f | cut -d / -f 2`
-   gfal-mkdir {srm}
-   echo "gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
-   gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
-   if [ $? -ne 0 ]; then
-      echo "ERROR: file $ff not copied correctly ?"
-   else
+if [ $looperExitStatus -ne 0 ]; then
+   echo "Looper failed. Don't attempt to copy corrupted file remotely"
+else
+   export LD_LIBRARY_PATH=/usr/lib64:$LD_LIBRARY_PATH # Fabio's workaround to fix gfal-tools with CMSSW
+   for f in Loop/tree*/*.root
+   do
+      ff=`basename $f | cut -d . -f 1`
+      d=`echo $f | cut -d / -f 2`
+      gfal-mkdir {srm}
+      echo "gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
+      gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
+      if [ $? -ne 0 ]; then
+         echo "ERROR: remote copy failed for file $ff"
+      else
+         echo "remote copy succeeded"
+         rm Loop/$d/$ff.root
+      fi
+   done
+fi
+""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm='srm://t3se01.psi.ch'+remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")]) + dirCopy
+   elif remoteDir.startswith("/dpm/oeaw.ac.at"):
+      cpCmd="""echo 'sending root files to remote dir'
+copyExitStat=0
+if [ $looperExitStatus -ne 0 ]; then
+   echo "Looper failed. Don't attempt to copy corrupted file remotely"
+else
+   export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/lib64/dcap/ # Fabio's workaround to fix gfal-tools
+   #nEvents=`grep 'number of events processed' Loop/log.txt | sed 's/.* \([0-9][0-9]*\)$/\\1/'`
+   nEvents=`grep 'All Events' Loop/skimAnalyzerCount/SkimReport.txt | sed 's/.* All Events  *\([0-9][0-9]*\).*/\\1/'`
+   echo "Processed ${{nEvents}} according to Loop/log.txt"
+   for f in Loop/tree*/*.root
+   do
+      ff=`basename $f | cut -d . -f 1`
+      d=`echo $f | cut -d / -f 2`
+      #gfal-mkdir {srm}
+      #echo "gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_nEvents${{nEvents}}.root"
+      #gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_nEvents${{nEvents}}.root
+      echo "Trying 10x: lcg-cp -v file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_Chunk{idx}_nEvents${{nEvents}}.root"
+      icnt=0
+      fileCopyExitStatus=1
+      while [ $icnt -lt 10 ]
+      do
+        lcg-cp -v file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_Chunk{idx}_nEvents${{nEvents}}.root
+        if [ $? -eq 0 ]
+        then
+          fileCopyExitStatus=0
+          break
+        else
+          let icnt+=1
+          sleep 300
+        fi
+      done
       rm Loop/$d/$ff.root
-   fi
-done
-echo 'sending the logs back'  # will send also root files if copy failed
-cp -r Loop/* $LS_SUBCWD""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm='srm://t3se01.psi.ch'+remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")])
+      if [ $fileCopyExitStatus -ne 0 ]; then
+         echo "ERROR: remote copy failed for file $ff"
+         copyExitStat=$fileCopyExitStatus
+      else
+         echo "remote copy succeeded"
+      fi
+   done
+fi
+if (( ( $looperExitStatus == 0 ) && ( $copyExitStat == 0 ) ))
+then
+  exit 0
+else
+  exit 1
+fi
+""".format(idx=0 if not "_Chunk" in jobDir else jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm='srm://hephyse.oeaw.ac.at'+remoteDir+jobDir[jobDir.rfind("/"):max(0,jobDir.find("_Chunk"))]) + dirCopy
    elif remoteDir.startswith("/eos/cms/store"):
        cpCmd="""echo 'sending root files to remote dir'
-export LD_LIBRARY_PATH=/usr/lib64:$LD_LIBRARY_PATH # Fabio's workaround to fix gfal-tools with CMSSW
-for f in Loop/*ree*/*.root
-do
-   ff=`basename $f | cut -d . -f 1`
-   d=`echo $f | cut -d / -f 2`
-   /afs/cern.ch/project/eos/installation/0.3.15/bin/eos.select mkdir {srm}
-   echo "cmsStage /`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
-   cmsStage /`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
-   if [ $? -ne 0 ]; then
-      echo "ERROR: file $ff not copied correctly ?"
-   else
-      rm Loop/$d/$ff.root
-   fi
-done
-echo 'sending the logs back'  # will send also root files if copy failed
-cp -r Loop/* $LS_SUBCWD""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm=(remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")]).split("/eos/cms",1)[1])
+if [ $looperExitStatus -ne 0 ]; then
+   echo "Looper failed. Don't attempt to copy corrupted file remotely"
+else
+   for f in Loop/*ree*/*.root
+   do
+      ff=`basename $f | cut -d . -f 1`
+      d=`echo $f | cut -d / -f 2`
+      /afs/cern.ch/project/eos/installation/0.3.15/bin/eos.select mkdir {srm}
+      echo "cmsStage /`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
+      cmsStage /`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
+      if [ $? -ne 0 ]; then
+         echo "ERROR: remote copy failed for file $ff"
+      else
+         echo "remote copy succeeded"
+         rm Loop/$d/$ff.root
+      fi
+   done
+fi
+""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm=(remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")]).split("/eos/cms",1)[1]) + dirCopy
    else:
        print "choose location not supported yet: ", remoteDir
        print 'path must start with "/pnfs/psi.ch" or "/eos/cms/store"'
@@ -96,6 +157,7 @@ ls
 cd `find . -type d | grep /`
 echo 'running'
 python $CMSSW_BASE/src/CMGTools/RootTools/python/fwlite/Looper.py config.pck
+looperExitStatus=$?
 echo
 {copy}
 """.format(copy=cpCmd)
@@ -129,27 +191,38 @@ def batchScriptPSI( jobDir, remoteDir=''):
    cmssw_release = os.environ['CMSSW_BASE']
    VO_CMS_SW_DIR = "/swshare/cms"  # $VO_CMS_SW_DIR doesn't seem to work in the new SL6 t3wn
 
+
+   dirCopy = """echo 'sending the logs back'  # will send also root files if copy failed
+cp -r Loop/* $SUBMISIONDIR
+if [ $? -ne 0 ]; then
+   echo 'ERROR: problem copying job directory back'
+else
+   echo 'job directory copy succeeded'
+fi"""
    if remoteDir=='':
-       cpCmd="""echo 'sending the job directory back'
-cp -r Loop/* $SUBMISIONDIR"""
+       cpCmd=dirCopy
    elif remoteDir.startswith("/pnfs/psi.ch"):
        cpCmd="""echo 'sending root files to remote dir'
-export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/lib64/dcap/ # Fabio's workaround to fix gfal-tools
-for f in Loop/tree*/*.root
-do
-   ff=`basename $f | cut -d . -f 1`
-   d=`echo $f | cut -d / -f 2`
-   gfal-mkdir {srm}
-   echo "gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
-   gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
-   if [ $? -ne 0 ]; then
-      echo "ERROR: file $ff not copied correctly ?"
-   else
-      rm Loop/$d/$ff.root
-   fi
-done
-echo 'sending the logs back'
-cp -r Loop/* $SUBMISIONDIR""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm='srm://t3se01.psi.ch'+remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")])
+if [ $looperExitStatus -ne 0 ]; then
+   echo "Looper failed. Don't attempt to copy corrupted file remotely"
+else
+   export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/lib64/dcap/ # Fabio's workaround to fix gfal-tools
+   for f in Loop/tree*/*.root
+   do
+      ff=`basename $f | cut -d . -f 1`
+      d=`echo $f | cut -d / -f 2`
+      gfal-mkdir {srm}
+      echo "gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root"
+      gfal-copy file://`pwd`/Loop/$d/$ff.root {srm}/${{ff}}_{idx}.root
+      if [ $? -ne 0 ]; then
+         echo "ERROR: remote copy failed for file $ff"
+      else
+         echo "remote copy succeeded"
+         rm Loop/$d/$ff.root
+      fi
+   done
+fi
+""".format(idx=jobDir[jobDir.find("_Chunk")+6:].strip("/"), srm='srm://t3se01.psi.ch'+remoteDir+jobDir[jobDir.rfind("/"):jobDir.find("_Chunk")]) + dirCopy
    else:
        print "remote directory not supported yet: ", remoteDir
        print 'path must start with "/pnfs/psi.ch"'
@@ -204,6 +277,7 @@ cd `find . -type d | grep /`
 echo 'running'
 #python $CMSSW_BASE/src/CMGTools/RootTools/python/fwlite/Looper.py config.pck
 python {cmssw}/src/CMGTools/RootTools/python/fwlite/Looper.py config.pck
+looperExitStatus=$?
 echo
 {copy}
 ###########################################################################
